+++ test/integration/tests/617-unused-flag-cli/Main.hs
@@ -1,0 +1,8 @@
+import StackTest
+
+main :: IO ()
+main = do
+    stack ["build"]
+    stackErr ["build", "--flag", "foo:bar"]
+    stackErr ["build", "--flag", "files:bar"]
+    stack ["build", "--flag", "*:bar"]