+++ test/integration/tests/617-unused-flag-name-yaml/Main.hs
@@ -1,0 +1,4 @@
+import StackTest
+
+main :: IO ()
+main = stackErr ["build"]